refactor: 테스트 클래스 폴더 수정
