+++ src/test/java/lotto/model/domain/BonusNumberTest.java
@@ -1,0 +1,5 @@
+package lotto.model.domain;
+
+class BonusNumberTest {
+
+}
