+++ src/test/java/lotto/model/domain/LottoTicketsTest.java
@@ -1,11 +1,10 @@
-package lotto.model;
+package lotto.model.domain;
 
 import static lotto.constant.ErrorMessages.EMPTY_INPUT_ERROR;
 import static lotto.constant.ErrorMessages.INVALID_MONEY_ERROR;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import lotto.model.domain.LottoTickets;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
