+++ src/test/java/lotto/model/domain/WinningLottoTest.java
@@ -1,10 +1,7 @@
-package lotto.model;
+package lotto.model.domain;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import lotto.model.domain.BonusNumber;
-import lotto.model.domain.WinningLotto;
-import lotto.model.domain.WinningNumbers;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
